requirements.txt
fastapi[standard]==0.115.14
uvicorn[standard]==0.35.0
pydantic==2.11.7
jinja2==3.1.6
python-dotenv==1.1.1
google-genai==1.23.0
pymongo==4.13.2